requirements.txt
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
flower==2.0.1
pandas==2.1.4
azure-storage-blob==12.19.0